requirements.txt
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.0.76
dlib==19.24.2
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.0
scikit-image==0.21.0 